Build an early list of cities, counties, and both together.
Build lists with a unique identifier, hashing the city and state name.
This deals with a lot of political geography: most towns and townships
(which are key in NJ and PA, at least) merged into the cities, and the
coextensive city/counties in VA.

But there remain gaps: townships with identical names in different
counties of the same state (such as the 47 Jackson Township, IN
instantiations) are omitted so far. There are places (e.g., CT) where
some duplication has made it in: a town and city with the same name and
same population, only differing by the designate. I might also want to
eliminate jurisdictions with a population of zero.