Bring duplicated cities/towns into the record.
Label places with degenerate ids with county or place IDs, patching them
in.

This should be a high water mark for line count in the CSV files:

  36093   95969 1483554 cities.csv
  39195  102853 1604738 cities_and_counties.csv
   3144    6974  122819 counties.csv

From here, we want to deduplicate some city/town and city/county row
pairs, but this should encompass every city, town(ship), or county I'd
run into in practice.