Clean up a last 6 coterminous townships manually.
Most of these were "City township" entities in Illinois, but not all, so
I resorted to manual review of all duplication suspects (subsets
matching on state and population, of which some but not all have the
"town" substring). Reviewing all such suspects of population >=4k led to
six townships to clean up.

Note the fall in line count:

  36048   95851 1481632 cities.csv
  39082  102580 1600064 cities_and_counties.csv
   3144    6974  122819 counties.csv

Close #2.